Consolidate plugin publishing settings
