+++ plugin/build.gradle.kts
@@ -25,18 +25,20 @@
  */
 @file:Suppress("RemoveRedundantQualifierName") // To prevent IDEA replacing FQN imports.
 
-import io.spine.internal.gradle.IncrementGuard
+import io.spine.internal.dependency.Grpc
 import io.spine.internal.dependency.Protobuf
-
+import io.spine.internal.gradle.IncrementGuard
+import io.spine.internal.gradle.isSnapshot
+import io.spine.internal.gradle.publish.PublishingRepos.cloudArtifactRegistry
+import java.io.FileWriter
+import java.util.Properties
 import org.apache.tools.ant.filters.ReplaceTokens
 
 plugins {
     id(io.spine.internal.dependency.Protobuf.GradlePlugin.id)
-
+    `java-gradle-plugin`
     id("com.gradle.plugin-publish").version("0.18.0")
     id("com.github.johnrengelman.shadow").version("7.1.0")
-    `bootstrap-plugin`
-    `prepare-config-resources`
 }
 
 apply<IncrementGuard>()
@@ -93,6 +95,19 @@ sourceSets {
     }
 }
 
+val pluginName = "spineBootstrapPlugin"
+
+gradlePlugin {
+    plugins {
+        create(pluginName) {
+            id = "io.spine.bootstrap"
+            implementationClass = "io.spine.tools.gradle.bootstrap.BootstrapPlugin"
+            displayName = "Spine Bootstrap"
+            description = "Prepares a Gradle project for development on Spine."
+        }
+    }
+}
+
 pluginBundle {
     website = "https://spine.io/"
     vcsUrl = "https://github.com/SpineEventEngine/bootstrap.git"
@@ -107,7 +122,7 @@ pluginBundle {
     withDependencies { clear() }
 
     plugins {
-        named("spineBootstrapPlugin") {
+        named(pluginName) {
             version = bootstrapVersion
         }
     }
@@ -135,6 +150,83 @@ tasks.shadowJar {
 
 artifacts {
     archives(tasks.shadowJar)
+}
+
+val bootstrapDir = file("$buildDir/bootstrap")
+
+sourceSets.main {
+    resources.srcDir(bootstrapDir)
+}
+
+val versionSnapshot = file("$bootstrapDir/artifact-snapshot.properties")
+val configDir = file("$rootDir/config")
+
+val timeVersion: String by extra
+val coreJavaVersion: String by extra
+val webVersion: String by extra
+val gCloudVersion: String by extra
+
+/*
+  This task creates the `artifact-snapshot.properties` file which is later added to the classpath of
+  the Bootstrap plugin. The file contains versions, artifact notations, repositories, etc. used in
+  the Gradle scripts which should also be used in the runtime of the Bootstrap plugin.
+
+  The keys for the `artifact-snapshot.properties` file are duplicated in
+  the `io.spine.tools.gradle.config.ArtifactSnapshot` class, where the file is parsed.
+ */
+val writeDependencies by tasks.registering {
+    group = "Spine bootstrapping"
+
+    inputs.dir(configDir)
+    inputs.property("version", bootstrapVersion)
+    outputs.file(versionSnapshot)
+
+    doFirst {
+        bootstrapDir.mkdirs()
+        if (!versionSnapshot.exists()) {
+            versionSnapshot.createNewFile()
+        }
+    }
+
+    doLast {
+        val artifacts = Properties()
+
+        artifacts.setProperty("spine.version.base", baseVersion)
+        artifacts.setProperty("spine.version.time", timeVersion)
+        artifacts.setProperty("spine.version.core", coreJavaVersion)
+        artifacts.setProperty("spine.version.web", webVersion)
+        artifacts.setProperty("spine.version.gcloud", gCloudVersion)
+        artifacts.setProperty("protobuf.compiler", Protobuf.compiler)
+        artifacts.setProperty("protobuf.java", Protobuf.libs[0])
+        artifacts.setProperty("grpc.stub", Grpc.stub)
+        artifacts.setProperty("grpc.protobuf", Grpc.protobuf)
+        artifacts.setProperty(
+            "repository.spine.release",
+            cloudArtifactRegistry.releases
+        )
+        artifacts.setProperty(
+            "repository.spine.snapshot",
+            cloudArtifactRegistry.snapshots
+        )
+
+        FileWriter(versionSnapshot).use {
+            artifacts.store(it, "Dependencies and versions required by Spine.")
+        }
+    }
+}
+
+tasks.processResources {
+    dependsOn(writeDependencies)
+}
+
+/**
+ * Publish the plugin at Gradle Plugin Portal only if it's not a snapshot version.
+ *
+ * See rules for publishing in the Gradle documentation
+ * ["How do I publish my plugin to the Plugin Portal?"](https://plugins.gradle.org/docs/publish-plugin).
+ */
+val publishPlugins: Task by tasks.getting {
+    enabled = !bootstrapVersion.isSnapshot()
 }
 
 tasks.publish {
